dodano wyjatki dokonczyc

+++ src/main/java/pl/edu/pjatk/MPR_Projekt/Service/PiesekService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import pl.edu.pjatk.MPR_Projekt.Model.Piesek;
 import pl.edu.pjatk.MPR_Projekt.Repository.PiesekRepository;
+import pl.edu.pjatk.MPR_Projekt.exception.PiesekAlreadyExistException;
+import pl.edu.pjatk.MPR_Projekt.exception.PiesekNotFoundException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,34 +13,52 @@ import java.util.Optional;
 
 @Component
 public class PiesekService {
-private PiesekRepository piesekRepository;
+    private PiesekRepository piesekRepository;
     private StringUtilsService stringUtilsService;
     List<Piesek> piesekList = new ArrayList<>();
 
-@Autowired
-public PiesekService(PiesekRepository repository, StringUtilsService stringUtilsService) {
-    this.piesekRepository = repository;
+    @Autowired
+    public PiesekService(PiesekRepository repository, StringUtilsService stringUtilsService) {
+        this.piesekRepository = repository;
 
-    piesekRepository.save(new Piesek("brown", "Spike", 1));
-    piesekRepository.save(new Piesek("red", "Doggy", 2));
-    piesekRepository.save(new Piesek("orange", "Leo", 3));
+        piesekRepository.save(new Piesek("brown", "Spike", 1));
+        piesekRepository.save(new Piesek("red", "Doggy", 2));
+        piesekRepository.save(new Piesek("orange", "Leo", 3));
     }
 
-    public List<Piesek> getPiesekByName (String name) {
-    return piesekRepository.findByName(name);
+    public List<Piesek> getPiesekByName(String name) {
+        List<Piesek> piesek = this.piesekRepository.findByName(name);
+        if (piesek.isEmpty()) {
+            throw new PiesekNotFoundException();
+        }
+        return piesekRepository.findByName(name);
     }
 
 
     public List<Piesek> getPiesekList() {
+        List<Piesek> piesek = (List<Piesek>) this.piesekRepository.findAll();
+        if (piesek.isEmpty()) {
+            throw new PiesekNotFoundException();
+        }
         return (List<Piesek>) piesekRepository.findAll();
     }
 
     public void createPiesek(Piesek piesek) {
-    piesekRepository.save(piesek);
+        piesek.obliczIdentyfikator();
+        List<Piesek> pieski = this.piesekRepository.findByIdentyfikator(piesek.getIdentyfikator());
+        if (!pieski.isEmpty()) {
+            throw new PiesekAlreadyExistException();
+        }
+        //sprawdz czy pola sa nie puste
+        piesekRepository.save(piesek);
     }
 
     public void deletePiesekById(int id) {
-piesekRepository.deleteById(id);
+        Optional<Piesek> piesek = piesekRepository.findById(id);
+        if (piesek.isEmpty()) {
+            throw new PiesekNotFoundException();
+        }
+        piesekRepository.deleteById(id);
     }
 
 //    Optional<Piesek> piesek = piesekRepository.findById(id);
@@ -49,19 +69,23 @@ piesekRepository.deleteById(id);
 //    }
 
     public void removePiesekByName(String name) {
-    List <Piesek> pieski = piesekRepository.findByName(name);
-    if (pieski.isEmpty()) {
-        throw new IllegalArgumentException("Piesek o nazwie" + name + " nie istnieje");
-    }
+        List<Piesek> pieski = piesekRepository.findByName(name);
+        if (pieski.isEmpty()) {
+            throw new PiesekNotFoundException();
+        }
         for (Piesek piesek : pieski) {
             piesekRepository.delete(piesek);
         }
     }
 
 
-public Optional<Piesek> getById(int id) {
-    return this.piesekRepository.findById(id);
-}
+    public Piesek getById(int id) {
+        Optional<Piesek> piesek = this.piesekRepository.findById(id);
+        if (piesek.isEmpty()) {
+            throw new PiesekNotFoundException();
+        }
+        return piesek.get();
+    }
 
 //    public void updatePiesek(int id, piesek ) {
 //       Optional<Piesek> existingPiesek = piesekRepository.findById(piesek.getId());
@@ -73,17 +97,17 @@ public Optional<Piesek> getById(int id) {
 //    }
 
 
-public void updatePiesek(int id, Piesek updatedPiesek) {
-    Optional<Piesek> existingPiesek = piesekRepository.findById(id);
-    if (existingPiesek.isPresent()) {
-        Piesek piesek = existingPiesek.get();
-        piesek.setName(updatedPiesek.getName());
-        piesek.setColor(updatedPiesek.getColor());
-        piesekRepository.save(piesek);
-    } else {
-        throw new RuntimeException("Piesek o ID " + id + " nie istnieje");
+    public void updatePiesek(int id, Piesek updatedPiesek) {
+        Optional<Piesek> existingPiesek = piesekRepository.findById(id);
+        if (existingPiesek.isPresent()) {
+            Piesek piesek = existingPiesek.get();
+            piesek.setName(updatedPiesek.getName());
+            piesek.setColor(updatedPiesek.getColor());
+            piesekRepository.save(piesek);
+        } else {
+            throw new PiesekNotFoundException();
+        }
     }
-}
 
 }
